refactor: adding IOutputItem interface and Output namespace
