+++ src/Output/IOutputItem.php
@@ -1,0 +1,10 @@
+<?php
+/**
+ * @author Victor Pryazhnikov <victor@pryazhnikov.com>
+ */
+namespace ReviewCombiner\Output;
+
+interface IOutputItem
+{
+    public function toOutputString(): string;
+}
